+++ api/server/services/auth.service.js
@@ -3,26 +3,19 @@ const Token = require('../../models/schema/tokenSchema');
 const sendEmail = require('../../utils/sendEmail');
 const crypto = require('crypto');
 const bcrypt = require('bcrypt');
-const DebugControl = require('../../utils/debug.js');
 const { registerSchema } = require('../../strategies/validators');
 const migrateDataToFirstUser = require('../../utils/migrateDataToFirstUser');
-
-function log({ title, parameters }) {
-  DebugControl.log.functionName(title);
-  DebugControl.log.parameters(parameters);
-}
 
 const isProduction = process.env.NODE_ENV === 'production';
 const clientUrl = isProduction ? process.env.CLIENT_URL_PROD : process.env.CLIENT_URL_DEV;
 
-const loginUser = async (user) => {
-  // const refreshToken = req.user.generateRefreshToken();
-  const dbUser = await User.findById(user._id);
-  //todo: save refresh token
-
-  return dbUser;
-};
-
+/**
+ * Logout user
+ *
+ * @param {Object} user 
+ * @param {*} refreshToken 
+ * @returns 
+ */
 const logoutUser = async (user, refreshToken) => {
   try {
     const userFound = await User.findById(user._id);
@@ -33,27 +26,29 @@ const logoutUser = async (user, refreshToken) => {
     }
 
     await userFound.save();
-    //res.clearCookie('refreshToken', COOKIE_OPTIONS);
-    // removeTokenCookie(res);
+
     return { status: 200, message: 'Logout successful' };
   } catch (err) {
     return { status: 500, message: err.message };
   }
 }
 
+/**
+ * Register a new user
+ *
+ * @param {Object} user <email, password, name, username>
+ * @returns 
+ */
 const registerUser = async (user) => {
-  let response = {};
   const { error } = registerSchema.validate(user);
   if (error) {
-    log({
-      title: 'Route: register - Joi Validation Error',
-      parameters: [
-        { name: 'Request params:', value: user },
-        { name: 'Validation error:', value: error.details }
-      ]
-    });
-    response = { status: 422, message: error.details[0].message };
-    return response;
+    console.info(
+      'Route: register - Joi Validation Error',
+      { name: 'Request params:', value: user },
+      { name: 'Validation error:', value: error.details }
+    );
+
+    return { status: 422, message: error.details[0].message };
   }
 
   const { email, password, name, username } = user;
@@ -62,61 +57,58 @@ const registerUser = async (user) => {
     const existingUser = await User.findOne({ email });
 
     if (existingUser) {
-      log({
-        title: 'Register User - Email in use',
-        parameters: [
-          { name: 'Request params:', value: user },
-          { name: 'Existing user:', value: existingUser }
-        ]
-      });
-      response = { status: 422, message: 'Email is in use' };
-      return response;
+      console.info(
+        'Register User - Email in use',
+        { name: 'Request params:', value: user },
+        { name: 'Existing user:', value: existingUser }
+      );
+      
+      // Sleep for 1 second
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+
+      // TODO: We should change the process to always email and be generic is signup works or fails (user enum)
+      return { status: 500, message: 'Something went wrong' };
     }
 
     //determine if this is the first registered user (not counting anonymous_user)
     const isFirstRegisteredUser = (await User.countDocuments({})) === 0;
 
-    try {
-      const newUser = await new User({
-        provider: 'local',
-        email,
-        password,
-        username,
-        name,
-        avatar: null,
-        role: isFirstRegisteredUser ? 'ADMIN' : 'USER'
-      });
+    const newUser = await new User({
+      provider: 'local',
+      email,
+      password,
+      username,
+      name,
+      avatar: null,
+      role: isFirstRegisteredUser ? 'ADMIN' : 'USER'
+    });
 
-      // todo: implement refresh token
-      // const refreshToken = newUser.generateRefreshToken();
-      // newUser.refreshToken.push({ refreshToken });
-      bcrypt.genSalt(10, (err, salt) => {
-        bcrypt.hash(newUser.password, salt, (errh, hash) => {
-          if (err) {
-            console.log(err);
-          }
-          // set pasword to hash
-          newUser.password = hash;
-          newUser.save();
-        });
-      });
-      console.log('newUser', newUser);
-      if (isFirstRegisteredUser) {
-        migrateDataToFirstUser(newUser);
-        // console.log(migrate);
-      }
-      response = { status: 200, user: newUser };
-      return response;
-    } catch (err) {
-      response = { status: 500, message: err.message };
-      return response;
+    // todo: implement refresh token
+
+    // Perform synchronously to prevent a race condition
+    const salt = await bcrypt.genSalt(10);
+    const hash = await bcrypt.hash(newUser.password, salt);
+    newUser.password = hash;
+    newUser.save();
+
+    // Legacy: This should be removed in a future release
+    if (isFirstRegisteredUser) {
+      migrateDataToFirstUser(newUser);
     }
+
+    return { status: 200, user: newUser };
+  
   } catch (err) {
-    response = { status: 500, message: err.message };
-    return response;
+    return { status: 500, message: err?.message || 'Something went wrong' };
   }
 };
-
+ 
+/**
+ * Request password reset
+ *
+ * @param {String} email 
+ * @returns 
+ */
 const requestPasswordReset = async (email) => {
   const user = await User.findOne({ email });
   if (!user) {
@@ -149,6 +141,14 @@ const requestPasswordReset = async (email) => {
   return { link };
 };
 
+/**
+ * Reset Password
+ *
+ * @param {*} userId 
+ * @param {String} token 
+ * @param {String} password 
+ * @returns 
+ */
 const resetPassword = async (userId, token, password) => {
   let passwordResetToken = await Token.findOne({ userId });
 
@@ -170,7 +170,7 @@ const resetPassword = async (userId, token, password) => {
 
   sendEmail(
     user.email,
-    'Password Reset Successfnodeully',
+    'Password Reset Successfully',
     {
       name: user.name
     },
@@ -183,9 +183,7 @@ const resetPassword = async (userId, token, password) => {
 };
 
 module.exports = {
-  // signup,
   registerUser,
-  loginUser,
   logoutUser,
   requestPasswordReset,
   resetPassword
